requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
python-dotenv
SQLAlchemy
Werkzeug
mariadb
requests-cache
